refactor: refactor correctColor files
