+++ src/correctColor/utils/formatData.ts
@@ -1,9 +1,8 @@
 import { colord, extend, RgbColor } from 'colord';
 import labPlugin from 'colord/plugins/lab';
 
-import { getRegressionVariables } from '../../correctColor';
-
-import { ColorCard } from './referenceColorCard';
+import { ColorCard } from '../__tests__/testUtils/referenceColorCard';
+import { getRegressionVariables } from '../correctColor';
 
 extend([labPlugin]);
