+++ src/correctColor/utils/getImageColors.ts
@@ -1,6 +1,6 @@
 import { RgbColor } from 'colord';
 
-import { IJS } from '../../../IJS';
+import { IJS } from '../../IJS';
 
 /**
  * Extract the colors of an image in order to use them for color correction. Should be used on small images only (smaller than 10x10 pixels), because it is these colors that will be used in the model (MLR).
